+++ CHANGES.md
@@ -2,11 +2,10 @@
 
 ## Version 2.0.0 ##
 
-Date: 2021-04-23
+Date: 2021-04-27
 
 - Code cleaning.
 - Remove storage abstractions.
-- Add custom `with-open` macro.
 
 
 ## Version 1.2.0 ##